testing travis, renamed server.js to index.js
