+++ .travis.yml
@@ -1,3 +1,3 @@
 language: node_js
 node_js:
-  - "stable"
+  - 0.12
