+++ package.json
@@ -1,11 +1,10 @@
 {
   "name": "acit2910",
   "version": "1.0.0",
-  "main": "server.js",
+  "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js",
-	"test": "mocha"
+    "test": "mocha",
+    "start": "node server.js"
   },
   "repository": {
     "type": "git",
@@ -28,6 +27,7 @@
   },
   "description": "",
   "devDependencies": {
+    "jshint": "^2.9.4",
     "mocha": "^3.3.0"
   }
 }
